Moving towards graphs being specified by configuration. All display items will have a class object to add the values to the display.
Changed configuration to use a stripped down version of the config
structure used in my other projects.  VarTab implements a serializable
hierarchical table of variables with values.  Getting a value from the
table an optionally perform macro replacement (from the table) or 'eval'
the results to produce an object of some type.  All items are stored as
strings or atomic types so serialization will always be possible.